Add a SolrDocument
