+++ lib/dpul_collections/indexing_pipeline/figgy/hydration_cache_entry.ex
@@ -1,9 +1,6 @@
 defmodule DpulCollections.IndexingPipeline.Figgy.HydrationCacheEntry do
-  alias DpulCollections.IndexingPipeline.Figgy.HydrationCacheEntry
-  alias DpulCollections.IndexingPipeline.Figgy
   use Ecto.Schema
   import Ecto.Changeset
-  require Logger
 
   schema "figgy_hydration_cache_entries" do
     field :data, :map
@@ -37,25 +34,5 @@ defmodule DpulCollections.IndexingPipeline.Figgy.HydrationCacheEntry do
       :source_cache_order,
       :source_cache_order_record_id
     ])
-  end
-
-  @spec to_solr_document(%__MODULE__{}) :: %{}
-  def to_solr_document(cache_entry = %HydrationCacheEntry{}) do
-    cache_entry
-    |> HydrationCacheEntry.to_combined_figgy_resource()
-    |> Figgy.CombinedFiggyResource.to_solr_document()
-  end
-
-  @spec to_combined_figgy_resource(%HydrationCacheEntry{}) :: %Figgy.CombinedFiggyResource{}
-  def to_combined_figgy_resource(%__MODULE__{
-        data: data,
-        related_data: related_data,
-        related_ids: related_ids
-      }) do
-    %Figgy.CombinedFiggyResource{
-      resource: data,
-      related_data: related_data,
-      related_ids: related_ids
-    }
   end
 end
