+++ test/dpul_collections/indexing_pipeline/figgy/resource_test.exs
@@ -3,16 +3,10 @@ defmodule DpulCollections.IndexingPipeline.Figgy.ResourceTest do
   alias DpulCollections.IndexingPipeline
   alias DpulCollections.IndexingPipeline.Figgy
 
-  describe "converting to solr" do
-    test "it's possible to convert a folder to a solr document without the pipeline" do
-      folder = IndexingPipeline.get_figgy_resource!("be12221a-6461-4aae-a6c6-c1defc8717dd")
-
-      doc =
-        folder
-        |> Figgy.Resource.to_combined()
-        |> Figgy.CombinedFiggyResource.to_solr_document()
-
-      assert doc[:format_txt_sort] == ["Ephemera"]
+  describe "populate_virtual/1" do
+    test "returns resource unchanged when metadata has no visibility or state" do
+      resource = IndexingPipeline.get_figgy_resource!("f99af4de-fed4-4baa-82b1-6e857b230306")
+      assert Figgy.Resource.populate_virtual(resource) == resource
     end
   end
